closes #217

+++ car/src/main/java/pt/ulisboa/tecnico/softeng/car/domain/Car.java
@@ -1,7 +1,7 @@
 package pt.ulisboa.tecnico.softeng.car.domain;
 
 public class Car extends Vehicle {
-	public Car(String plate, int kilometers, RentACar rentACar, float price) {
+	public Car(String plate, int kilometers, RentACar rentACar, int price) {
 		super(plate, kilometers, rentACar, price);
 	}
 }
